fix: move pricing to settings submodule
